Edit Ineration with Gun
Add the ability to enter a socket slot when a gun is picked up.
Adde the ability to pick up a gun and swap the gun you're holding when the slot is full.
Add ability to cancel gun physics when picked up and apply physics when dropped.
Activate the gun's gravity(I don't think I turn off it. why?)