Error fixed -  It was an error caused by Unreal.
Fixed the issue where character objects hit by bullets were not affected.
Obviously all bones already have the same collision options set. Collision enabled (query and physics).
It seems to be an error in Unreal.
Just in case, I selected the option again and saved it, and it is affected normally. What the fuck is this!
my 3 days!!!