Invecntory function Update
Inventory Feature Updates
When picking up an item, the item is saved by checking the maximum number of stacks of the picked up item, the number of items currently in stock, and whether there is empty space.

Applies the default image to the item slot.
Make the item image and amount appear only when picking up an item from the item slot.